fix: address remaining PR #822 review feedback
tool-args-recovery-boundary.tsx:
- Add effective-cap note near MAX_RECOVERY_ATTEMPTS constant: stale-state
  reads mean the effective cap is MAX + 1 (4 not 3) due to async setState
- Add explanatory comment in render() permanent-fallback branch explaining
  why no new recovery is scheduled when both hasArgsTextError and
  recoveryAttempt >= MAX_RECOVERY_ATTEMPTS are true

web-search-ui.tsx, code-interpreter-ui.tsx:
- Add missing newline at EOF

lib/utils/image-validation.ts (NEW):
- Extract shared isSafePlotData, SAFE_PLOT_MIME_TYPES, and
  SAFE_IMAGE_MIME_TYPES to prevent allowlist drift between
  code-interpreter-ui.tsx and connector-tool-ui.tsx

code-interpreter-ui.tsx:
- Replace inline isSafePlotData with import from shared module

connector-tool-ui.tsx:
- Replace inline SAFE_IMAGE_MIME_TYPES with import from shared module

tests/components/tool-args-recovery-boundary.test.tsx:
- Add transient-null test: verifies boundary renders null in the 50ms
  recovery window before the timer fires (toBeEmptyDOMElement assertion)
